로그인, 로그아웃, 댓글 아이콘 보이는 기능 리팩토링
